requirements.txt
Flask
Flask_Bcrypt
mysql-connector-python
flask-login